Updates to statistics.
Fixes the following bugs:
- Cloning ConfigComponent did not copy configuration for stats enabled with enableAll* functions
- Some of the statistic configuration data was not being serialized
- Python function for setting stat load level on a component would segfault
- Recent stat changes made explicitly enabled stats erroneously check stat load level
- All stats for a component would get the same params as the first enabled stat
- Stat params were not always propagating, making stop-at, start-at, rate, etc. not work

Other changes:
- TXT and console stat output now inherit from same class and match output
  - TXT output turns on rank and simtime output by default
  - Console output turns off rank and simtime output by default (existing behavior)
- Reset on output now resets the entire statistic
  - Count for event based output separated from count counted in output

+++ src/sst/core/model/python/pymodel_comp.cc
@@ -374,7 +374,7 @@ static PyObject*
 compSetStatisticLoadLevel(PyObject* self, PyObject* args)
 {
     int              argOK             = 0;
-    uint8_t          loadLevel         = STATISTICLOADLEVELUNINITIALIZED;
+    int              loadLevel         = STATISTICLOADLEVELUNINITIALIZED;
     ConfigComponent* c                 = getComp(self);
     int              apply_to_children = 0;
 
@@ -387,6 +387,7 @@ compSetStatisticLoadLevel(PyObject* self, PyObject* args)
     else {
         return nullptr;
     }
+
     return SST_ConvertToPythonLong(0);
 }
